feat: provide config.xml snippet in readme for ui colors
Copying this line into the config file will be a lot less annoying
than manually entering the rgb values for every element in the settings
GUI.

I actually generated these using a whiskers template, but I wasn't sure
how that should be integrated into the repo. So I didn't include it in
this commit.

Would like to see if these instructions are clear and work for others.