feat: use tera to template the ui theme instructions
Allows control over the UI colors from a template with a single output
file.

Link to the instructions in the README. This avoids making the entire
README a whiskers template.

Also update the Just recipes and the .gitattributes for a new generated
file

squashing and force pushing because I don't like Github's auto web
commit history spam