feat: update bash highlighting to match styleguide
Brings the bash syntax theme up to date with style guide.

The lexer is not very advanced, and doesn't support many contexts.

This will not look nearly as good a vscode or neovim with a proper
language server or tree-sitter.

But it's probably the best Npp can do with the builtin language support.

Relates to #17