feat: add missing and new global style groups
Adds all the style groups to match the default (stylers.xml) available
in the most recent version of Npp (v8.7.7).

I have also re-ordered the list to match the default style for clarity.

Includes these groups:
- Multi-selected text color
- Multi-edit carets color
- Bookmark margin
- Change History Margin
- Fold Margin
- Incremental highlight all ( Search > Incremental Search [Ctrl+Alt+I] )
- Active tab unfoced indicator ( I can't identify what this actually
  changes)
- Active tab text
- Inactive tabs (foreground and background)
- Document map (view > document map)

In order for tab styles to work, I enabled the attribute
"darkTabUseTheme" in the config.xml modification. There does not seem to
be a way to set the active tab background color in the theme, it will
use the "Active" custom UI color (currently `mantle`)